requirements.txt
# AI World Tracker - Production Dependencies

# Core Data Collection
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11  # Fixed for Python 3.13+ (cgi module removed)
arxiv>=2.1.3  # Updated for feedparser compatibility
# legacy-cgi only needed for Python 3.13+ (cgi module was removed)
# Install manually if using Python 3.13+: pip install legacy-cgi>=2.6.1

# Date/Time Handling
python-dateutil>=2.8.0
pytz>=2023.3

# Data Visualization
matplotlib>=3.7.0

# Optional AI Features
# openai>=1.6.0

# Development Dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0

pyyaml
